Make marker formatting independent of operation order
The marker display code assumes that all versions are normalized, in that all trailing zeroes are stripped. This is not the case for star versions, where the trailing zeroes (before the `.*`) are semantically relevant. This would cause path dependent-behavior where we would get a different marker string depending on whether a version with or without a trailing zero was added to the cache first. This would cause `maker_normalization_a` but `maker_normalization_b` to fail.